+++ uploadpagina/session.inc.php
@@ -1,0 +1,9 @@
+<?php
+
+session_start();
+
+if(!isset($_SESSION['username'])){
+
+    header("location:login.php");
+    exit;
+}
